feat: Análise completa com Dart Analyzer, flutter_lints e dart_code_metrics - Redução de 1118 issues (38.8%) com flutter clean + pub get + dart fix --apply

+++ lib/modules/products/presentation/providers/products_state_provider.dart
@@ -646,7 +646,7 @@ class ProductsListNotifier extends StateNotifier<ProductsListState> {
 
       }
 
-    } catch (e.toDouble() ?? 0.0) {
+    } catch (e() ?? 0.0) {
  // ignore: argument_type_not_assignable
 
       // ignore: argument_type_not_assignable
